requirements.txt
# Core ML and image processing
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Adversarial libraries
foolbox>=3.3.3
adversarial-robustness-toolbox>=1.15.0

# CLI and utilities
click>=8.1.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-image>=0.20.0

# For Fawkes integration
requests>=2.31.0 